+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [unreleased]
 
+ * Default Login: Improved the style and added another config property *pageTitle* for the title in
+   the header (Default is the server name).
+
 ## 2.0.0-alpha.3 (February 7, 2022)
 
  * Admin Toolbar: The Portal App selection shows now the i18n *title* instead of the App name and the i18n *description*